day 9 part i
a little bit slow but good enough for this puzzle

This is a combination of 2 commits.
This is the 1st commit message:

day 9 part i passes tests but I think it is gonna be really slow

let's see

This is the commit message #2:

ok it took 4 seconds on my raspberry pi 4

good enough but let's see how they multiply this in part ii!